+++ src/main/java/kitchenpos/product/application/ProductService.java
@@ -1,11 +1,15 @@
 package kitchenpos.product.application;
 
 import kitchenpos.product.domain.Product;
+import kitchenpos.product.dto.ProductRequest;
+import kitchenpos.product.dto.ProductResponse;
 import kitchenpos.product.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+
+import static java.util.stream.Collectors.toList;
 
 @Service
 public class ProductService {
@@ -16,12 +20,15 @@ public class ProductService {
     }
 
     @Transactional
-    public Product create(final Product product) {
-        return productRepository.save(product);
+    public ProductResponse create(final ProductRequest product) {
+        Product savedProduct = productRepository.save(product.toEntity());
+        return new ProductResponse(savedProduct);
     }
 
     @Transactional(readOnly = true)
-    public List<Product> list() {
-        return productRepository.findAll();
+    public List<ProductResponse> list() {
+        return productRepository.findAll().stream()
+                .map(ProductResponse::new)
+                .collect(toList());
     }
 }
