+++ src/main/java/kitchenpos/product/dto/ProductRequest.java
@@ -1,0 +1,36 @@
+package kitchenpos.product.dto;
+
+import kitchenpos.product.domain.Product;
+
+public class ProductRequest {
+    private String name;
+    private long price;
+
+    public ProductRequest() {
+    }
+
+    public ProductRequest(String name, long price) {
+        this.name = name;
+        this.price = price;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public long getPrice() {
+        return price;
+    }
+
+    public void setPrice(long price) {
+        this.price = price;
+    }
+
+    public Product toEntity() {
+        return new Product(name, price);
+    }
+}
