+++ src/main/java/kitchenpos/product/dto/ProductResponse.java
@@ -1,0 +1,39 @@
+package kitchenpos.product.dto;
+
+import kitchenpos.product.domain.Product;
+
+public class ProductResponse {
+    private Long id;
+    private String name;
+    private long price;
+
+    public ProductResponse(Product created) {
+        this.id = created.getId();
+        this.name = created.getName();
+        this.price = created.getPrice().getValue();
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public void setId(Long id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public long getPrice() {
+        return price;
+    }
+
+    public void setPrice(long price) {
+        this.price = price;
+    }
+}
