+++ src/main/java/kitchenpos/product/ui/ProductRestController.java
@@ -1,7 +1,8 @@
 package kitchenpos.product.ui;
 
 import kitchenpos.product.application.ProductService;
-import kitchenpos.product.domain.Product;
+import kitchenpos.product.dto.ProductRequest;
+import kitchenpos.product.dto.ProductResponse;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -10,7 +11,6 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.net.URI;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @RestController
 public class ProductRestController {
@@ -22,78 +22,17 @@ public class ProductRestController {
 
     @PostMapping("/api/products")
     public ResponseEntity<ProductResponse> create(@RequestBody final ProductRequest request) {
-        final Product created = productService.create(request.toProduct());
+        final ProductResponse created = productService.create(request);
         final URI uri = URI.create("/api/products/" + created.getId());
         return ResponseEntity.created(uri)
-                .body(new ProductResponse(created))
+                .body(created)
                 ;
     }
 
     @GetMapping("/api/products")
     public ResponseEntity<List<ProductResponse>> list() {
         return ResponseEntity.ok()
-                .body(productService.list().stream().map(ProductResponse::new).collect(Collectors.toList()))
+                .body(productService.list())
                 ;
-    }
-
-    static class ProductRequest {
-        private String name;
-        private long price;
-
-        public String getName() {
-            return name;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
-
-        public long getPrice() {
-            return price;
-        }
-
-        public void setPrice(long price) {
-            this.price = price;
-        }
-
-        public Product toProduct() {
-            return new Product(name, price);
-        }
-    }
-
-    static class ProductResponse {
-        private Long id;
-        private String name;
-        private long price;
-
-        public ProductResponse(Product created) {
-            this.id = created.getId();
-            this.name = created.getName();
-            this.price = created.getPrice().getValue();
-        }
-
-        public Long getId() {
-            return id;
-        }
-
-        public void setId(Long id) {
-            this.id = id;
-        }
-
-        public String getName() {
-            return name;
-        }
-
-        public void setName(String name) {
-            this.name = name;
-        }
-
-        public long getPrice() {
-            return price;
-        }
-
-        public void setPrice(long price) {
-            this.price = price;
-        }
     }
 }
