+++ src/test/java/kitchenpos/product/application/ProductServiceTest.java
@@ -1,6 +1,8 @@
 package kitchenpos.product.application;
 
 import kitchenpos.product.domain.Product;
+import kitchenpos.product.dto.ProductRequest;
+import kitchenpos.product.dto.ProductResponse;
 import kitchenpos.product.repository.ProductRepository;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -8,7 +10,10 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 
 @DisplayName("상품 서비스 관련")
 @SpringBootTest
@@ -22,13 +27,14 @@ class ProductServiceTest {
     @Test
     void create() {
         // given
-        Product given = new Product("짜장면", 6000);
+        ProductRequest given = new ProductRequest("짜장면", 6000);
+        when(productRepository.save(any(Product.class))).thenReturn(given.toEntity());
 
         // when
-        productService.create(given);
+        ProductResponse actual = productService.create(given);
 
         // then
-        verify(productRepository).save(given);
+        assertThat(actual.getName()).isEqualTo(given.getName());
     }
 
     @DisplayName("상품의 목록을 조회할 수 있다")
